Improve support for wrapping go-errors with New,
The go-errors package puts a huge emphasis on the original stack trace. Even
if you wrap an error the original stack trace is the one that is shown. I've
never seen this before but it is very useful since it is always the original
stack trace that shows you where your code have problems.

But there is one usage pattern that causes the original stack trace to
disappear and that is if you use errors.New(...) on an error that already has
a stack trace. We did that by mistake in our code and since this only shows
the outer stack trace it made trouble shooting harder.

This commit makes all levels of such a re-newed error visible while keeping
the emphasis on the original stack trace by using a Java like formatting of
the stack trace but reversing the order.

This is an example where an error has been New'ed three times:

error_test.go:103: *errors.errorString foo
	/Users/john.doe/go/src/github.com/go-errors/errors/error_test.go:92 (0x10f8d39)
		TestMultipleNew.func1: return New("foo")
	/Users/john.doe/go/src/github.com/go-errors/errors/error_test.go:96 (0x10f8d86)
		TestMultipleNew.func2: return New(inner())
	/Users/john.doe/go/src/github.com/go-errors/errors/error_test.go:99 (0x10f6244)
		TestMultipleNew: err := New(outer())
	/usr/local/Cellar/go/1.9.4/libexec/src/testing/testing.go:746 (0x10b8060)
		tRunner: fn(t)
	/usr/local/Cellar/go/1.9.4/libexec/src/runtime/asm_amd64.s:2337 (0x1056b61)
		goexit: BYTE	$0x90	// NOP

	Which caused: *errors.Error foo
	/Users/john.doe/go/src/github.com/go-errors/errors/error_test.go:96 (0x10f8d9c)
		TestMultipleNew.func2: return New(inner())
	/Users/john.doe/go/src/github.com/go-errors/errors/error_test.go:99 (0x10f6244)
		TestMultipleNew: err := New(outer())
	/usr/local/Cellar/go/1.9.4/libexec/src/testing/testing.go:746 (0x10b8060)
		tRunner: fn(t)
	/usr/local/Cellar/go/1.9.4/libexec/src/runtime/asm_amd64.s:2337 (0x1056b61)
		goexit: BYTE	$0x90	// NOP

	Which caused: *errors.Error foo
	/Users/john.doe/go/src/github.com/go-errors/errors/error_test.go:99 (0x10f625a)
		TestMultipleNew: err := New(outer())
	/usr/local/Cellar/go/1.9.4/libexec/src/testing/testing.go:746 (0x10b8060)
		tRunner: fn(t)
	/usr/local/Cellar/go/1.9.4/libexec/src/runtime/asm_amd64.s:2337 (0x1056b61)
		goexit: BYTE	$0x90	// NOP